+++ src/main/java/app/entities/products/materials/planks/Rafter.java
@@ -7,11 +7,11 @@ import java.util.List;
 
 public class Rafter extends Plank {
 
-    public Rafter(String name, String description, float costPrice, float salesPrice, float width, String unit, List<Integer> preCutsLengths, int height) {
+    public Rafter(String name, String description, float costPrice, float salesPrice, float width, String unit, List<Integer> preCutsLengths, float height) {
         super(name, description, costPrice, salesPrice, width, unit, preCutsLengths, height);
     }
 
-    public Rafter(int itemId, String name, String description, float costPrice, float salesPrice, String unit, float width, int height) {
+    public Rafter(int itemId, String name, String description, float costPrice, float salesPrice, String unit, float width, float height) {
         super(itemId, name, description, costPrice, salesPrice, unit, width, height);
     }
 
@@ -21,7 +21,7 @@ public class Rafter extends Plank {
         ps.setString(2, this.getDescription());
         ps.setString(3, this.getUnit());
         ps.setFloat(4, this.getWidth());
-        ps.setInt( 5, this.getHeight());
+        ps.setFloat( 5, this.getHeight());
         ps.setString(6, this.getClass().getSimpleName());
         ps.setNull(7, Types.NUMERIC);
         ps.setNull(8, Types.NUMERIC);
